Move (Read|Write)Iterations to their own header files
