+++ src/Series.cpp
@@ -24,6 +24,7 @@
 #include "openPMD/IO/AbstractIOHandler.hpp"
 #include "openPMD/IO/AbstractIOHandlerHelper.hpp"
 #include "openPMD/IO/Format.hpp"
+#include "openPMD/ReadIterations.hpp"
 #include "openPMD/Series.hpp"
 #include "openPMD/version.hpp"
 
@@ -1309,195 +1310,5 @@ namespace
                 return [](std::string const &) -> std::tuple<bool, int> { return std::tuple<bool, int>{false, 0}; };
         }
     }
-} // namespace [anonymous]
-
-SeriesIterator::SeriesIterator() : m_series()
-{
-}
-
-SeriesIterator::SeriesIterator( Series series )
-    : m_series( std::move( series ) )
-{
-    auto it = series.get().iterations.begin();
-    if( it == series.get().iterations.end() )
-    {
-        *this = end();
-        return;
-    }
-    else
-    {
-        auto status = it->second.beginStep();
-        if( status == AdvanceStatus::OVER )
-        {
-            *this = end();
-            return;
-        }
-        it->second.setStepStatus( StepStatus::DuringStep );
-    }
-    m_currentIteration = it->first;
-}
-
-SeriesIterator &
-SeriesIterator::operator++()
-{
-    if( !m_series.has_value() )
-    {
-        *this = end();
-        return *this;
-    }
-    Series & series = m_series.get();
-    auto & iterations = series.iterations;
-    auto & currentIteration = iterations[ m_currentIteration ];
-    if( !currentIteration.closed() )
-    {
-        currentIteration.close();
-    }
-    switch( series.iterationEncoding() )
-    {
-        using IE = IterationEncoding;
-        case IE::groupBased:
-        {
-            // since we are in group-based iteration layout, it does not
-            // matter which iteration we begin a step upon
-            AdvanceStatus status = currentIteration.beginStep();
-            if( status == AdvanceStatus::OVER )
-            {
-                *this = end();
-                return *this;
-            }
-            currentIteration.setStepStatus( StepStatus::DuringStep );
-            break;
-        }
-        default:
-            break;
-    }
-    auto it = iterations.find( m_currentIteration );
-    auto itEnd = iterations.end();
-    if( it == itEnd )
-    {
-        *this = end();
-        return *this;
-    }
-    ++it;
-    if( it == itEnd )
-    {
-        *this = end();
-        return *this;
-    }
-    m_currentIteration = it->first;
-    switch( series.iterationEncoding() )
-    {
-        using IE = IterationEncoding;
-        case IE::fileBased:
-        {
-            auto & iteration = series.iterations[ m_currentIteration ];
-            AdvanceStatus status = iteration.beginStep();
-            if( status == AdvanceStatus::OVER )
-            {
-                *this = end();
-                return *this;
-            }
-            iteration.setStepStatus( StepStatus::DuringStep );
-            break;
-        }
-        default:
-            break;
-    }
-    return *this;
-}
-
-IndexedIteration
-SeriesIterator::operator*()
-{
-    return IndexedIteration(
-        m_series.get().iterations[ m_currentIteration ], m_currentIteration );
-}
-
-bool
-SeriesIterator::operator==( SeriesIterator const & other ) const
-{
-    return this->m_currentIteration == other.m_currentIteration &&
-        this->m_series.has_value() == other.m_series.has_value();
-}
-
-bool
-SeriesIterator::operator!=( SeriesIterator const & other ) const
-{
-    return !operator==( other );
-}
-
-SeriesIterator
-SeriesIterator::end()
-{
-    return {};
-}
-
-ReadIterations::ReadIterations( Series series )
-    : m_series( std::move( series ) )
-{
-}
-
-ReadIterations::iterator_t
-ReadIterations::begin()
-{
-    return iterator_t{ m_series };
-}
-
-ReadIterations::iterator_t
-ReadIterations::end()
-{
-    return SeriesIterator::end();
-}
-
-WriteIterations::SharedResources::SharedResources( iterations_t _iterations )
-    : iterations( std::move( _iterations ) )
-{
-}
-
-WriteIterations::SharedResources::~SharedResources()
-{
-    if( currentlyOpen.has_value() )
-    {
-        auto lastIterationIndex = currentlyOpen.get();
-        auto & lastIteration = iterations.at( lastIterationIndex );
-        if( !lastIteration.closed() )
-        {
-            lastIteration.close();
-        }
-    }
-}
-
-WriteIterations::WriteIterations( iterations_t iterations )
-    : shared{ std::make_shared< SharedResources >( std::move( iterations ) ) }
-{
-}
-
-WriteIterations::mapped_type &
-WriteIterations::operator[]( key_type const & key )
-{
-    // make a copy
-    // explicit cast so MSVC can figure out how to do it correctly
-    return operator[]( static_cast< key_type && >( key_type{ key } ) );
-}
-WriteIterations::mapped_type &
-WriteIterations::operator[]( key_type && key )
-{
-    if( shared->currentlyOpen.has_value() )
-    {
-        auto lastIterationIndex = shared->currentlyOpen.get();
-        auto & lastIteration = shared->iterations.at( lastIterationIndex );
-        if( lastIterationIndex != key && !lastIteration.closed() )
-        {
-            lastIteration.close();
-        }
-    }
-    shared->currentlyOpen = key;
-    auto & res = shared->iterations[ std::move( key ) ];
-    if( res.getStepStatus() == StepStatus::NoStep )
-    {
-        res.beginStep();
-        res.setStepStatus( StepStatus::DuringStep );
-    }
-    return res;
-}
-} // namespace openPMD
+    } // namespace [anonymous]
+    } // namespace openPMD
